split form_factor into universial.py and kang.py methods
