+++ sparrowpy/form_factor/universial.py
@@ -9,10 +9,48 @@ except ImportError:
     numba = None
     prange = range
 
+def patch2patch_ff_universal(patches_points: np.ndarray,
+                             patches_normals: np.ndarray,
+                             patches_areas: np.ndarray,
+                             visible_patches:np.ndarray):
+    """Calculate the form factors between patches (universal method).
 
-#/////////////////////////////////////////////////////////////////////////////////////#
-#######################################################################################
-### patch-to-patch
+    This method computes form factors between polygonal patches via numerical
+    integration. Support is only guaranteed for convex polygons.
+    Because this function relies on numpy, all patches must have
+    the same number of sides.
+
+    Parameters
+    ----------
+    patches_points : np.ndarray
+        vertices of all patches of shape (n_patches, n_vertices, 3)
+    patches_normals : np.ndarray
+        normal vectors of all patches of shape (n_patches, 3)
+    patches_areas : np.ndarray
+        areas of all patches of shape (n_patches)
+    visible_patches : np.ndarray
+        index list of all visible patches combinations (n_combinations, 2)
+
+    Returns
+    -------
+    form_factors : np.ndarray
+        form factors between all patches of shape (n_patches, n_patches)
+        note that just i_source < i_receiver are calculated since
+        patches_areas[i] * ff[i, j] = patches_areas[j] * ff[j, i]
+
+    """
+    n_patches = patches_areas.shape[0]
+    form_factors = np.zeros((n_patches, n_patches))
+
+    for visID in prange(visible_patches.shape[0]):
+        i = int(visible_patches[visID, 0])
+        j = int(visible_patches[visID, 1])
+        form_factors[i,j] = calc_form_factor(
+                    patches_points[i], patches_normals[i], patches_areas[i],
+                    patches_points[j], patches_normals[j])
+
+    return form_factors
+
 def calc_form_factor(source_pts: np.ndarray, source_normal: np.ndarray,
                      source_area: np.ndarray, receiver_pts: np.ndarray,
                      receiver_normal: np.ndarray,
@@ -41,370 +79,242 @@ def calc_form_factor(source_pts: np.ndarray, source_normal: np.ndarray,
 
     Returns
     -------
-    out: float
+    form_factor: float
         form factor
 
     """
     if geom._coincidence_check(receiver_pts, source_pts):
-        out = nusselt_integration(
+        form_factor = integration.nusselt_integration(
                     patch_i=source_pts, patch_i_normal=source_normal,
                     patch_j=receiver_pts, patch_j_normal=receiver_normal,
                     nsamples=64)
     else:
-        out = stokes_integration(patch_i=source_pts, patch_j=receiver_pts,
-                                 patch_i_area=source_area,  approx_order=4)
+        form_factor = integration.stokes_integration(patch_i=source_pts,
+                                             patch_j=receiver_pts,
+                                             patch_i_area=source_area,
+                                             approx_order=4)
 
-    return out
+    return form_factor
 
-#######################################################################################
-def load_stokes_entries(
-    i_bpoints: np.ndarray, j_bpoints: np.ndarray) -> np.ndarray:
-    """Load all the stokes form function values between two patches.
+## accumulation of propagation effects (air attenuation, scattering)
 
-    Parameters
-    ----------
-    i_bpoints: np.ndarray
-        list of points in patch i boundary (n_boundary_points_i , 3)
+def _form_factors_with_directivity(
+        visibility_matrix, form_factors, n_bins, patches_center,
+        patches_area, air_attenuation,
+        absorption, absorption_index, patch_to_wall_ids,
+        scattering, scattering_index, sources, receivers):
+    """Calculate the form factors with directivity."""
+    n_patches = patches_center.shape[0]
+    form_factors_tilde = np.zeros((n_patches, n_patches, n_patches, n_bins))
+    # loop over previous patches, current and next patch
 
-    j_bpoints: np.ndarray
-        list of points in patch j boundary (n_boundary_points_j , 3)
+    for ii in prange(n_patches**3):
+        h = ii % n_patches
+        i = int(ii/n_patches) % n_patches
+        j = int(ii/n_patches**2) % n_patches
+        visible_hi = visibility_matrix[
+            h, i] if h < i else visibility_matrix[i, h]
+        visible_ij = visibility_matrix[
+            i, j] if i < j else visibility_matrix[j, i]
+        if visible_hi and visible_ij:
+            difference_receiver = patches_center[i]-patches_center[j]
+            wall_id_i = int(patch_to_wall_ids[i])
+            difference_receiver /= np.linalg.norm(difference_receiver)
+            ff = form_factors[i, j] if i<j else (form_factors[j, i]
+                                                 *patches_area[i]/patches_area[j])
 
-    Returns
-    -------
-    form_mat: np.ndarray
-        f function value matrix (n_boundary_points_i , n_boundary_points_j)
+            distance = np.linalg.norm(difference_receiver)
+            form_factors_tilde[h, i, j, :] = ff
+            if air_attenuation is not None:
+                form_factors_tilde[h, i, j, :] = form_factors_tilde[
+                    h, i, j, :] * np.exp(-air_attenuation * distance)
 
-    """
-    form_mat = np.zeros((len(i_bpoints) , len(j_bpoints)))
+            if scattering is not None:
+                scattering_factor = get_scattering_data(
+                    patches_center[h], patches_center[i], patches_center[j],
+                    sources, receivers, wall_id_i,
+                    scattering, scattering_index)
+                form_factors_tilde[h, i, j, :] = form_factors_tilde[
+                    h, i, j, :] * scattering_factor
 
-    for i in prange(i_bpoints.shape[0]):
-        for j in prange(j_bpoints.shape[0]):
-            form_mat[i][j] = np.log(np.linalg.norm(i_bpoints[i]-j_bpoints[j]))
+            if absorption is not None:
+                source_wall_idx = absorption_index[wall_id_i]
+                form_factors_tilde[h, i, j, :] = form_factors_tilde[
+                    h, i, j, :] * (1-absorption[source_wall_idx])
 
-    return form_mat
-
-def stokes_integration(
-    patch_i: np.ndarray, patch_j: np.ndarray, patch_i_area: float,
-    approx_order=4) -> float:
-    """Calculate an estimation of the form factor between two patches.
-
-    Computationally integrates a modified form function over
-    the boundaries of both patches.
-    The modified form function follows Stokes' theorem.
-
-    The modified form function integral is calculated using a
-    polynomial approximation based on sampled values.
-
-    Parameters
-    ----------
-    patch_i : np.ndarray
-        vertex coordinates of patch i (n_vertices, 3)
-
-    patch_i_area: float
-       area of patch i
-
-    patch_j : np.ndarray
-        vertex coordinates of patch j (n_vertices, 3)
-
-    source_area: float
-        area of the source patch
-
-    approx_order: int
-        polynomial order of the form function integration estimation
-
-    Returns
-    -------
-    float
-    form factor between two patches
-
-    """
-    i_bpoints, i_conn = integration._sample_boundary_regular(patch_i,
-                                                         npoints=approx_order+1)
-    j_bpoints, j_conn = integration._sample_boundary_regular(patch_j,
-                                                         npoints=approx_order+1)
-
-    subsecj = np.zeros((j_conn.shape[1]))
-    subseci = np.zeros((i_conn.shape[1]))
-    form_mat = np.zeros((i_bpoints.shape[0],j_bpoints.shape[0]))
-
-    # first compute and store form function sample values
-    form_mat = load_stokes_entries(i_bpoints, j_bpoints)
-
-    # double polynomial integration (per dimension (x,y,z))
-    outer_integral = 0
-    inner_integral = np.zeros((len(i_bpoints),len(j_bpoints[0])))
-
-    for dim in range(len(j_bpoints[0])): # for each dimension
-        # integrate form function over each point on patch i boundary
-
-        for i in range(len(i_bpoints)):   # for each point in patch i boundary
-            for segj in j_conn:     # for each segment segj in patch j boundary
-
-                xj = j_bpoints[segj][:,dim]
-
-                if np.abs(xj[-1]-xj[0])>1e-3:
-                    for k in range(len(segj)):
-                        subsecj[k] = form_mat[i][segj[k]]
-
-                    # compute polynomial coefficients
-                    quadfactors = integration._poly_estimation_Lagrange(x=xj,
-                                                                   y=subsecj)
-                    # analytical integration of the approx polynomials
-                    inner_integral[i][dim] += integration._poly_integration(
-                                                        c=quadfactors,x=xj)
+    return form_factors_tilde
 
 
-        # integrate previously computed integral over patch i
-        for segi in i_conn:   # for each segment segi in patch i boundary
+def _form_factors_with_directivity_dim(
+        visibility_matrix, form_factors, n_bins, patches_center,
+        patches_area,
+        air_attenuation,
+        patch_to_wall_ids,
+        scattering, scattering_index, sources, receivers):
+    """Calculate the form factors with directivity."""
+    n_patches = patches_center.shape[0]
+    n_directions = receivers.shape[1] if receivers is not None else 1
+    form_factors_tilde = np.zeros((n_patches, n_patches, n_directions, n_bins))
+    # loop over previous patches, current and next patch
 
-            xi = i_bpoints[segi][:,dim]
+    for ii in prange(n_patches**2):
+        i = ii % n_patches
+        j = int(ii/n_patches) % n_patches
+        visible_ij = visibility_matrix[
+            i, j] if i < j else visibility_matrix[j, i]
+        if visible_ij:
+            difference_receiver = patches_center[i]-patches_center[j]
+            wall_id_i = int(patch_to_wall_ids[i])
+            difference_receiver /= np.linalg.norm(difference_receiver)
+            ff = form_factors[i, j] if i<j else (form_factors[j, i]
+                                                 *patches_area[j]/patches_area[i])
 
-            if np.abs(xi[-1]-xi[0])>1e-3:
-                for k in range(len(segi)):
-                    subseci[k] = inner_integral[segi[k]][dim]
-                quadfactors = integration._poly_estimation_Lagrange(x=xi,
-                                                                y=subseci)
-                outer_integral += integration._poly_integration(c=quadfactors,
-                                                                x=xi)
+            distance = np.linalg.norm(difference_receiver)
+            form_factors_tilde[i, j, :, :] = ff
+            if air_attenuation is not None:
+                form_factors_tilde[i, j, :, :] = form_factors_tilde[
+                    i, j, :, :] * np.exp(-air_attenuation * distance)
 
-    return np.abs(outer_integral/(2*np.pi*patch_i_area))
+            if scattering is not None:
+                scattering_factor = get_scattering_data_source(
+                    patches_center[i], patches_center[j],
+                    sources, wall_id_i,
+                    scattering, scattering_index)
+                form_factors_tilde[i, j, :, :] = form_factors_tilde[
+                    i, j, :, :] * scattering_factor
 
-def nusselt_analog(surf_origin, surf_normal,
-                   patch_points, patch_normal) -> float:
-    """Calculate the Nusselt analog for a single point.
+    return form_factors_tilde
 
-    Projects a given receiver patch onto a hemisphere centered around a point
-    on a source patch surface.
-    The hemispherical projection is then projected onto the source patch plane.
-    The area of this projection relative to the unit circle area is the
-    differential form factor between the two patches.
+
+## scattering data access
+
+def get_scattering_data_receiver_index(
+        pos_i:np.ndarray, pos_j:np.ndarray,
+        receivers:np.ndarray, wall_id_i:np.ndarray,
+        ):
+    """Get scattering data depending on previous, current and next position.
 
     Parameters
     ----------
-    surf_origin : np.ndarray
-        point on source patch for differential form factor evaluation (3,)
-        (global origin)
-
-    surf_normal : np.ndarray
-        normal of source patch (3,)
-
-    patch_points : np.ndarray
-        vertex coordinates of the receiver patch (n_vertices, 3)
-
-    patch_normal: np.ndarray
-        normal of receiver patch (3,)
+    pos_i : np.ndarray
+        current position of shape (3)
+    pos_j : np.ndarray
+        next position of shape (3)
+    receivers : np.ndarray
+        receiver directions of all walls of shape (n_walls, n_receivers, 3)
+    wall_id_i : np.ndarray
+        current wall id to get write directional data
 
     Returns
     -------
-    Nusselt analog factor
-    (differential form factor)
+    scattering_factor: float
+        scattering factor from directivity
 
     """
-    boundary_points, connectivity = integration._sample_boundary_regular(
-                                                            patch_points,
-                                                            npoints=3)
+    n_patches = pos_i.shape[0] if pos_i.ndim > 1 else 1
+    receiver_idx = np.empty((n_patches), dtype=np.int64)
 
-    hand = np.sign(np.dot(
-                np.cross(patch_points[1]-patch_points[0],
-                         patch_points[2]-patch_points[1]), patch_normal) )
-
-    curved_area = 0
-
-    sphPts = np.empty_like( boundary_points )
-    projPts = np.empty_like( boundary_points )
-    plnPts = np.empty( shape=(len(boundary_points),2) )
-
-    for ii in prange(len(boundary_points)):
-        # patch j points projected on the hemisphere
-        sphPts[ii] = ( (boundary_points[ii]-surf_origin) /
-                        np.linalg.norm(boundary_points[ii]-surf_origin) )
-
-    rotmat = geom._rotation_matrix(n_in=surf_normal)
-
-    for ii in prange(len(sphPts)):
-        # points on the hemisphere projected onto patch plane
-        plnPts[ii,:] = geom._matrix_vector_product(matrix=rotmat,
-                                                      vector=sphPts[ii])[:-1]
-        projPts[ii,:-1] = plnPts[ii,:]
-        projPts[ii,-1] = 0.
+    for i in range(n_patches):
+        difference_receiver = pos_i[i]-pos_j
+        difference_receiver /= np.linalg.norm(
+            difference_receiver)
+        receiver_idx[i] = np.argmin(np.sum(
+            (receivers[wall_id_i[i], :]-difference_receiver)**2, axis=-1),
+            axis=-1)
 
 
-    big_poly = geom._polygon_area(projPts[0::2])
-
-    segmt=np.empty_like(connectivity[0])
-
-    leftseg=np.empty((3,2))
-    rightseg=np.empty((3,2))
-
-    for jj in prange(connectivity.shape[0]):
-
-        segmt = connectivity[jj]
-
-        if (np.linalg.norm(np.cross(projPts[segmt[-1]],projPts[segmt[0]]))
-                                                                    > 1e-6):
-
-            # if the points on the segment span less than 90 degrees
-            if np.dot( plnPts[segmt[-1]], plnPts[segmt[0]] ) >= 1e-6:
-                curved_area += integration._area_under_curve(plnPts[segmt],order=2)
-
-            # if points span over 90º, additional sampling is required
-            else:
-                mpoint = ( sphPts[segmt[0]] +
-                          (sphPts[segmt[-1]] - sphPts[segmt[0]]) / 2 )
-
-                # midpoint on the arc projected on the hemisphere
-                marc = mpoint/np.linalg.norm(mpoint)
-                a = sphPts[segmt[0]] + (marc - sphPts[segmt[0]]) / 2
-                b = marc + (sphPts[segmt[-1]] - marc) / 2
-
-                mpoint = geom._matrix_vector_product(matrix=rotmat,
-                                                        vector=mpoint)[:-1]
-                marc = geom._matrix_vector_product(matrix=rotmat,
-                                                      vector=marc)[:-1]
-                a = a/np.linalg.norm(a)
-                a = geom._matrix_vector_product(matrix=rotmat,vector=a)[:-1]
-
-                b = b/np.linalg.norm(b)
-                b = geom._matrix_vector_product(matrix=rotmat,vector=b)[:-1]
-
-                linArea = (np.linalg.norm(plnPts[segmt[-1]]-plnPts[segmt[0]])
-                                              * np.linalg.norm(mpoint-marc)/2)
-
-                leftseg[0] = plnPts[segmt[0]]
-                leftseg[1] = a
-                leftseg[2] = marc
-
-                rightseg[0] = marc
-                rightseg[1] = b
-                rightseg[2] = plnPts[segmt[-1]]
+    return receiver_idx
 
 
-                left =  integration._area_under_curve(leftseg, order=2)
-                right = integration._area_under_curve(rightseg, order=2)
-                curved_area += (linArea * np.sign(left) + left + right)
-
-    return big_poly + hand*curved_area
-
-def nusselt_integration(patch_i: np.ndarray, patch_j: np.ndarray,
-                        patch_i_normal: np.ndarray, patch_j_normal: np.ndarray,
-                        nsamples=2, random=False) -> float:
-    """Estimate the form factor based on the Nusselt analogue.
-
-    Integrates the differential form factor (Nusselt analogue output)
-    over the surface of the source patch
+def get_scattering_data(
+        pos_h:np.ndarray, pos_i:np.ndarray, pos_j:np.ndarray,
+        sources:np.ndarray, receivers:np.ndarray, wall_id_i:np.ndarray,
+        scattering:np.ndarray, scattering_index:np.ndarray):
+    """Get scattering data depending on previous, current and next position.
 
     Parameters
     ----------
-    patch_i: np.ndarray
-        vertex coordinates of the source patch
-
-    patch_j: np.ndarray
-        vertex coordinates of the receiver patch
-
-    patch_i_normal: np.ndarray
-        source patch normal (3,)
-
-    patch_j_normal: np.ndarray
-        receiver patch normal (3,)
-
-    patch_i_area: float
-        source patch area
-
-    patch_j_area: float
-        receiver patch area
-
-    nsamples: int
-        number of receiver surface samples for integration
-
-    random: bool
-        determines the distribution of the samples on patch_i surface
-        if True, the samples are randomly distributed in a uniform way
-        if False, a regular sampling of the surface is performed
+    pos_h : np.ndarray
+        previous position of shape (3)
+    pos_i : np.ndarray
+        current position of shape (3)
+    pos_j : np.ndarray
+        next position of shape (3)
+    sources : np.ndarray
+        source directions of all walls of shape (n_walls, n_sources, 3)
+    receivers : np.ndarray
+        receiver directions of all walls of shape (n_walls, n_receivers, 3)
+    wall_id_i : np.ndarray
+        current wall id to get write directional data
+    scattering : np.ndarray
+        scattering data of shape (n_scattering, n_sources, n_receivers, n_bins)
+    scattering_index : np.ndarray
+        index of the scattering data of shape (n_walls)
 
     Returns
     -------
-    out: float
-        form factor between patches i and j
+    scattering_factor: float
+        scattering factor from directivity
 
     """
-    if random:
-        p0_array = integration._surf_sample_random(patch_i,nsamples)
-    else:
-        p0_array = integration._surf_sample_regulargrid(patch_i,nsamples)
+    difference_source = pos_h-pos_i
+    difference_receiver = pos_i-pos_j
 
-    out = 0
-
-    for i in prange(p0_array.shape[0]):
-        out += nusselt_analog( surf_origin=p0_array[i],
-                               surf_normal=patch_i_normal,
-                               patch_points=patch_j,
-                               patch_normal=patch_j_normal )
-
-    out *= 1 / ( np.pi * len(p0_array) )
-
-    return out
+    difference_source /= np.linalg.norm(difference_source)
+    difference_receiver /= np.linalg.norm(difference_receiver)
+    source_idx = np.argmin(np.sum(
+        (sources[wall_id_i, :, :]-difference_source)**2, axis=-1))
+    receiver_idx = np.argmin(np.sum(
+        (receivers[wall_id_i, :]-difference_receiver)**2, axis=-1))
+    return scattering[scattering_index[wall_id_i],
+        source_idx, receiver_idx, :]
 
 
-#/////////////////////////////////////////////////////////////////////////////////////#
-#######################################################################################
-### point-to-patch and patch-to-point
-def pt_solution(point: np.ndarray, patch_points: np.ndarray, mode='source'):
-    """Calculate the geometric factor between a point and a patch.
-
-    applies a modified version of the Nusselt analogue,
-    transformed for a -point- source rather than differential surface element.
+def get_scattering_data_source(
+        pos_h:np.ndarray, pos_i:np.ndarray,
+        sources:np.ndarray, wall_id_i:np.ndarray,
+        scattering:np.ndarray, scattering_index:np.ndarray):
+    """Get scattering data depending on previous, current position.
 
     Parameters
     ----------
-    point: np.ndarray
-        source or receiver point
-
-    patch_points: np.ndarray
-        vertex coordinates of the patch
-
-    mode: string
-        determines if point is acting as a source ('source')
-        or as a receiver ('receiver')
+    pos_h : np.ndarray
+        previous position of shape (3)
+    pos_i : np.ndarray
+        current position of shape (3)
+    sources : np.ndarray
+        source directions of all walls of shape (n_walls, n_sources, 3)
+    wall_id_i : np.ndarray
+        current wall id to get write directional data
+    scattering : np.ndarray
+        scattering data of shape (n_scattering, n_sources, n_receivers, n_bins)
+    scattering_index : np.ndarray
+        index of the scattering data of shape (n_walls)
 
     Returns
     -------
-    geometric factor
+    scattering_factor: float
+        scattering factor from directivity
 
     """
-    if mode == 'receiver':
-        source_area = geom._polygon_area(patch_points)
-    elif mode == 'source':
-        source_area = 4
-
-    npoints = len(patch_points)
-
-    interior_angle_sum = 0
-
-    patch_onsphere = np.zeros_like(patch_points)
-
-    for i in range(npoints):
-        patch_onsphere[i]= ( (patch_points[i]-point) /
-                              np.linalg.norm(patch_points[i]-point) )
-
-    for i in range(npoints):
-
-        v0 = geom._sphere_tangent_vector(patch_onsphere[i],
-                                              patch_onsphere[(i-1)%npoints])
-        v1 = geom._sphere_tangent_vector(patch_onsphere[i],
-                                              patch_onsphere[(i+1)%npoints])
-
-        interior_angle_sum += np.arccos(np.dot(v0,v1))
-
-    factor = interior_angle_sum - (len(patch_points)-2)*np.pi
-
-    return factor / (np.pi*source_area)
+    difference_source = pos_h-pos_i
+    difference_source /= np.linalg.norm(difference_source)
+    source_idx = np.argmin(np.sum(
+        (sources[wall_id_i, :, :]-difference_source)**2, axis=-1))
+    return scattering[scattering_index[wall_id_i], source_idx]
 
 
 if numba is not None:
+    patch2patch_ff_universal = numba.njit(parallel=True)(
+        patch2patch_ff_universal)
     calc_form_factor = numba.njit()(calc_form_factor)
-    pt_solution = numba.njit(parallel=True)(pt_solution)
-    nusselt_integration = numba.njit(parallel=False)(nusselt_integration)
-    stokes_integration = numba.njit(parallel=False)(stokes_integration)
-    nusselt_analog = numba.njit(parallel=False)(nusselt_analog)
-    load_stokes_entries = numba.njit(parallel=True)(load_stokes_entries)
+    _form_factors_with_directivity_dim = numba.njit(parallel=True)(
+        _form_factors_with_directivity_dim)
+    _form_factors_with_directivity = numba.njit(parallel=True)(
+        _form_factors_with_directivity)
+    get_scattering_data_receiver_index = numba.njit()(
+        get_scattering_data_receiver_index)
+    get_scattering_data = numba.njit()(get_scattering_data)
+    get_scattering_data_source = numba.njit()(get_scattering_data_source)
+
